% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ntc.R
\name{append_ntc_pair_annotations}
\alias{append_ntc_pair_annotations}
\title{Append stored non-targeting-control annotations}
\usage{
append_ntc_pair_annotations(results, gi_obj)
}
\description{
Append stored non-targeting-control annotations
}
\keyword{internal}
