% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CeRberus-package.R
\docType{package}
\name{CeRberus-package}
\alias{CeRberus}
\alias{CeRberus-package}
\title{CeRberus: Aggregate Guide-Level CRISPR Genetic Interaction Scores}
\description{
Aggregates guide-level CRISPR genetic interaction scores to gene-pair-level scores using limma-based models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thorohde/CeRberus}
}

}
\author{
\strong{Maintainer}: Thomas Rohde \email{thorohde@uni-bonn.de} (\href{https://orcid.org/0009-0004-5159-4635}{ORCID})

Authors:
\itemize{
  \item Thomas Rohde \email{thorohde@uni-bonn.de} (\href{https://orcid.org/0009-0004-5159-4635}{ORCID})
}

}
\keyword{internal}
