% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ScreenDesign.R
\name{extract-ScreenDesign}
\alias{extract-ScreenDesign}
\alias{$,ScreenDesign-method}
\alias{[[,ScreenDesign,character-method}
\title{Access fields in an internal screen design}
\usage{
\S4method{$}{ScreenDesign}(x, name)

\S4method{[[}{ScreenDesign,character}(x, i)
}
\arguments{
\item{x}{A \code{ScreenDesign} object.}

\item{name}{Character scalar naming a design field.}

\item{i}{Character scalar naming a design field.}
}
\value{
The value stored in the selected field.
}
\description{
These methods provide read-only list-like access to fields stored in an
internal \code{\link[=ScreenDesign-class]{ScreenDesign}} object.
}
\keyword{internal}
