% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_ScreenBase.R
\docType{class}
\name{FixedPairScreen-class}
\alias{FixedPairScreen-class}
\title{Fixed-pair CRISPR screen}
\description{
S4 class for screens where each guide-pair maps to a fixed gene pair.
}
\keyword{internal}
