% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ntc.R
\name{label_ntc_pairs}
\alias{label_ntc_pairs}
\title{Label result pairs involving non-targeting controls}
\usage{
label_ntc_pairs(results, non_targeting_controls = NULL)
}
\description{
Label result pairs involving non-targeting controls
}
\keyword{internal}
