% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_ScreenBase.R
\docType{class}
\name{MultiplexScreen-class}
\alias{MultiplexScreen-class}
\title{Multiplex CRISPR screen}
\description{
S4 class for multiplex screens where query genes are modelled against a
library-gene axis.
}
\keyword{internal}
