% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_gene_pairs.R
\name{sort_gene_pairs}
\alias{sort_gene_pairs}
\title{Alphabetically sorts gene pairs.}
\usage{
sort_gene_pairs(g1, g2, sep = ";", pairs, pair_sep = ";", invert = FALSE)
}
\arguments{
\item{g1}{Character vector of gene names}

\item{g2}{Character vector of gene names}

\item{sep}{Separator used to concatenate the gene pairs}

\item{pairs}{Vector of gene pair names, separated by the pair_sep separator}

\item{pair_sep}{Separator used when pairs is provided}

\item{invert}{inversts the sorting of gene pairs. If false, the output will be "GeneA;GeneB". If true: "GeneB;GeneA"}
}
\description{
After pairwise sorting, \code{sort_gene_pairs} concatenates two vectors of gene names into one.
}
\examples{
genes1 <- c("RB1", "NOTCH1", "TTN", "MSH2", "FANCD2")
genes2 <- c("RNF43", "NEK1", "HLTF", "REV3L", "PAPD7")

# Basic usage
sort_gene_pairs(genes1, genes2)

# Inverted sorting
sort_gene_pairs(genes1, genes2, invert = TRUE)

}
