% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ntc.R
\name{store_ntc_pair_annotations}
\alias{store_ntc_pair_annotations}
\title{Store non-targeting-control pair annotations}
\usage{
store_ntc_pair_annotations(gi_obj)
}
\description{
Store non-targeting-control pair annotations
}
\keyword{internal}
