% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_run.R
\name{full_run}
\alias{full_run}
\title{Run the Full GI Analysis Pipeline.}
\usage{
full_run(yaml_fpath, return_output = TRUE)
}
\arguments{
\item{yaml_fpath}{Path to a YAML instruction file.}

\item{return_output}{Binary that defines if the function should save}
}
\value{
Saves the GI scores and plots to the output directory defined in
the YAML file.
}
\description{
Run the Full GI Analysis Pipeline.
}
